More V2 migration updates, added C compilation support, .peko.h header support, and starting migrating new parser, simulator and codegen patterns

+++ crates/peko-core/src/asts/values.rs
@@ -34,7 +34,7 @@ impl Spanned for BooleanAST {
 /// based on the surrounding context.
 #[derive(Clone, new)]
 pub struct NumberAST {
-    pub value: PositionedValue<f32>,
+    pub value: PositionedValue<f64>,
     pub integer: bool,
 }
